refactor(utils): add getDiffInCalendarSeconds as ESM
