+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export { default as getDiffInMilliseconds } from './getDiffInMilliseconds.js';
+export { default as getDiffInCalendarSeconds } from './getDiffInCalendarSeconds.js';
